Made a linear and logistical regression models
Data set 2 seems to be giving us a hard time. I think this is a result of a few things like the type of features included in the set, the fact that there is no correlation between any features and the price, and there are a handful of features that are non-numerical data types. My models yield high error and low accuracy